requirements.txt
FastAPI
SQLAlchemy
pymysql
uvicorn
python-dotenv